vast/requirements.txt
# Data ingestion
vastdb>=1.0.0
kafka-python>=2.0.2
pyarrow>=14.0.0

# Diagnostic chat tool
anthropic>=0.40.0
trino>=0.330.0

# Web UI
flask>=3.0.0

# Failure simulator (direct PostgreSQL connection for config changes)
psycopg2-binary>=2.9.0

# Predictive maintenance (optional - statistical methods used if not available)
scikit-learn>=1.3.0
numpy>=1.24.0

# OpenTelemetry instrumentation
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0
opentelemetry-instrumentation-flask>=0.41b0
opentelemetry-instrumentation-requests>=0.41b0
